==> 00_pre-requisites.ipynb <==
# Pre-requisites
This document outlines how to set-up environment 
 - for [python](https://www.python.org/downloads/)
 - using [PyCharm](https://www.jetbrains.com/pycharm/download/) and [Jupyter Notebook](http://jupyter.org/install) as IDE
 - to run [IBKR TWS API]( https://interactivebrokers.github.io/) in C:\ root for Windows
 - on [TWS](https://www.interactivebrokers.com.hk/en/index.php?f=16042) or [IB Gateway](https://www.interactivebrokers.com.hk/en/index.php?f=16457)
 - with [IB_insync](https://rawgit.com/erdewit/ib_insync/master/docs/html/readme.html#) library
 - backed up into [git](https://git-scm.com/downloads)
 
 **Note:** The instructions are for Windows 10. The same process can be used for other OS - such as Ubuntu and MacOS.
 
## Notes to ready the software
  1. IBKR's TWS API should be in the root folder. This needs to be shared between Python and Jupyter.
    
  2. For PyCharm _Create New Project_ in on any empty folder (let us call it _IB_). It is important that you do this first.
    * Use the downloaded python (using the full path) as the interpreter of choice for PyCharm.
    * In _File_ -> _Settings_ -> _Project:IB_ add ib-insync packages. Other packages (such as pandas) can also be added on a need basis.
    * Add _TWS API_ as a Root Content with _File_ -> _Settings_ -> _Project:IB_ -> _Project Structure_ -> _Add Root Content_
    * Mark Directory as Source Root via _Project_ under _IB_ -> _source_ -> _pythonclient_ -> _ibapi_
    * Make a new Directory called _python_ in _IB_ and write your scripts there
      
  3. To get Jupyter recognize TWS API, go to the _C:\TWS API\source\pythonclient_ and run _python setup.py install_
    
  4. Set the API for TWS / IB Gateway to the appropriate _Socket Port_
    * For our example we will use IB Gateway's paper trading account with a Socket Port of 4002
    
  5. To update ib_insync to latest version (distribution), open anaconda and type *pip3 install -U ib_insync* in it. It will automatically update in *Users>User>Miniconda3>Lib>site-packages* the *ib_insync* and *ib_insync_0.x.xx.dist-info* folders
  
## Check the setup
Run the following:
### Importing

import os
os.getcwd() 

from ib_insync import *
util.startLoop() # Works only in notebooks, not in regular Python programs

### Connecting

ib=IB() # Instance of IB main class
ib.connect('127.0.0.1', 4004, clientId=1)    # Run this only on Paper Trades

### Test the connection

import ib_insync
print(ib_insync.__all__)

[v for v in ib.accountValues() if v.tag == 'NetLiquidationByCurrency' and v.currency == 'BASE']

# Output would be something like:
#[AccountValue(account='DUxxxxxx', tag='NetLiquidationByCurrency', value='1009909.26', currency='BASE', modelCode='')]

# %%time

contract = Stock('TSLA', 'SMART', 'USD')
cds = ib.reqContractDetails(contract)

cds

### Disconnect
To disconnect the session

ib.disconnect()
