==> 09_nse_place_put_trades.ipynb <==
# Program places orders

# STATUS: Incomplete

#***          Start ib_insync (run once)       *****
#___________________________________________________

from ib_insync import *
util.startLoop()
# ib=IB().connect('127.0.0.1', 7496, clientId=1) # kavi TWS live
ib=IB().connect('127.0.0.1', 4001, clientId=1) # kavi IBG live

import pandas as pd
import datetime

#******         Paths and variables         ****
#_______________________________________________

datapath = r'./zdata/'
base_rom = 20    # 90% RoM atleast

df_orders = pd.read_pickle(datapath+'df_orders.pkl')

# Filter for puts below strike threshold
df_orders = df_orders[df_orders.Strike <= df_orders.Threshold]

# Filter out the top Return on Margin trades
df_orders1 = df_orders[df_orders.RoM >= base_rom].reset_index(drop=True)
df_orders1 = df_orders1.sort_values('RoM', ascending=False)

# Compute margin for the trade
df_orders1['Margin'] = df_orders1.marginpct*df_orders1.Mlot*df_orders1.Strike

# Filter only the trades expiring in this month
# df = df_orders1[df_orders1.DTE == df_orders1.DTE.min()].reset_index(drop=True)
df = df_orders1

# sorted(df_orders1.ibSymbol.unique())
df.Expiry.min()

# total margin
df.Margin.sum()

len(df)

# Sort by Symbol for easy reference in Excel
df.sort_values(by='ibSymbol').to_excel(datapath+'temp.xlsx', index=None, freeze_panes=(1,1))

df.ibSymbol.unique()

#------------------------------------------------
#####    WARNING: THIS PLACES THE TRADE  #######
#________________________________________________

limitTrade = [ib.placeOrder(contract, order) 
              for contract, order in zip(df.Contract, df.Orders)]

#------------------------------------------------
#####         Code to cancel all trades    ############
#________________________________________________
cancel = ib.reqGlobalCancel()


limitTrade

# Put placed orders to a dataframe
df_orders_placed = pd.DataFrame(limitTrade)

# Write orders placed to file
x  =  datapath + datetime.datetime.now().strftime("%Y%m%d_%H%M%S") + "_orders_placed.xlsx"
writer = pd.ExcelWriter(x)
df_orders_placed.to_excel(writer, 'all-options', index=False, header=None)
writer.save()

ib.disconnect()

list(df_orders)

df.head()
